requirements.txt
pandas
numpy
scikit-learn
streamlit
llama-cpp-python==0.2.24

